requirements.txt
Pillow==10.0.0
Requests==2.31.0
selenium==4.12.0
webdriver_manager==4.0.0
